safe_spaces.py
For the necessary conversion from alphanumerical to 0-indexed coordinates in the first level i decided
to search for the letter's position in the unicode, since that was faster than using a dictionary as i
found out using timeit. This is inversed in level 3 using the chr() function

I had two approaches for the second level in my mind. One was to see the agents' locations as origins of a
spreading area that once it had made contact with another areas border would not expand further into that direction.
This would ensure that every coordinate's  distance to an agent would only have to be calculated once
(unless it's a border between areas) and would always be the minimum. 
The other approach just calculated every coordinate's distance to every agent and compared the minimum found distance
of that specicific coordinate to the greatest found distance on the entire grid. If it's equal to the greatest,
the coordinate is added to the safe_spots list. If it's larger than the greatest, the safe_spots list is emptied
and the greatest distance raised.
In the end i went with the second approach as i felt it was sufficient for a 10x10 grid and therefore a respectively
small dataset. With a growing matrix however the first approach might safe more time. Another problem I faced and was
not able to overcome was to code the "don't move in that direction anymore" without the algorithm checking for every
coordinate, since that is actually slower than just calculating the distance.